Constructing infrastructure
In this update, I created two objects for each of three basic layers
(the background, the player, the input box). One for view code and the
other for logic. There is also a comment referring to inertia, which we
plan on implementing when we get started with motion.
Minor edit: converted two const ints to an enum.